vsg 后端重写 M11q–M11t：D5 半浮点 albedo、D3 色彩空间契约（demo 是第一个实例）
- M11q §11.16cm 收益分析：把"附件内 16 级/6.5% 跳变"换算成用户看得见的量，D5 排第一
- M11r §11.16cn 两条会话句子的守卫用例（灯掉落 / 空矩形）
- M11s D5 G-buffer 的 albedo 附件 RGBA8 -> RGBA16F；端到端两光强区制测量说明收益有条件
  （x0.5 两种格式完全一样，x4 时 8 位 2 LSB/16 级 vs 16F 1 LSB/62 级）；引擎格式契约断言
- M11t D3 色彩空间契约写进四处 SDK 注释（Colorf/PixelFormat/RenderTarget/BuiltinShaders）
  + usage.md §3.9；demo 的六张 JPG 天空面 Rgba8Unorm -> Rgba8Srgb（视口带 mean 122.7 -> 80.8，
  门禁三条判据未动）；变异证明解码来自 vkFormatFor 那一个映射

证据：两棵树门禁 8 stage 全 ok cases=441 failed=0 vuid=0 hazard=0 skipped=0、doc symbols
145 unit(s)、应用 vuid=0 warnings=0（debug 与 release 逐字相同）；test_graphics 两棵树各 276；
设计日志 §11.16cm–§11.16cp。

+++ src/viz/graphics/src/RenderPipelineBuilder.cpp
@@ -84,7 +84,17 @@ intrusive_ptr<RenderTarget> RenderPipelineBuilder::defaultGbufferTarget(int widt
     }
     auto gbuffer = make_intrusive<RenderTarget>();
     gbuffer->setSize(width, height);
-    gbuffer->attachColor(RenderTarget::ColorFormat::RGBA8);   // att 0: albedo
+    // att 0: albedo - HALF FLOAT, not 8 bit (D5, wired 2026-09-25). What the extra bits buy was measured END
+    // TO END, because the display is 8 bit too: with a DIM light (x0.5, the demo's regime) the output is
+    // identical either way - the output's own quantisation is the limit (both: 1 LSB of error, 9 levels over a
+    // 0..0.06 ramp) - while with a BRIGHT one (x4) the 8-bit albedo's step reaches the eye multiplied: 2 LSB of
+    // error and 16 levels against 1 LSB and 62 levels for the half-float store (see ContentPassTest's
+    // `MeasureWhatAHalfFloatAlbedoWouldBuyEndToEnd`). So this is future-proofing for hosts that light with an
+    // intensity above 1, at +8.3 MB per 1080p G-buffer and one recompile of the offscreen pipelines (the shape
+    // is part of their key). Trade-off worth stating: the colour read-back packs RGBA8 only, so this attachment
+    // can no longer be read back as a target attachment - the previews copy it through a program, which is how
+    // the demo shows it.
+    gbuffer->attachColor(RenderTarget::ColorFormat::RGBA16F);
     gbuffer->attachColor(RenderTarget::ColorFormat::RGBA16F); // att 1: view normal (+ shininess)
     gbuffer->attachColor(RenderTarget::ColorFormat::RGBA8);   // att 2: specular
     gbuffer->attachColor(RenderTarget::ColorFormat::RGBA16F); // att 3: view position
